Working Roadrunner auto for duck side on red and blue
